Display benchmark results as they become available
The old DoBenchmark() would fill in a list of the results, and Main()
would just dump the list. This makes the user wait for a long time
staring a blank screen that just asks them to be patient.

Since the benchmarking code is not dependent on anything that maybe
happening with the console I/O, DoBenchmark() has been change to return
an IEnumerable<string> and the result of each hashing benchmarking
result is yielded as they become available. This gives the user more
immediate feedback.