Use LINQ to search hashes to benchmark
The code for finding the hash algorithms to benchmark was using the
procedural style of filling a list to be processed instead of the
modern C# functional style of consuming an IEnumerable. It was also
harder to discern what was actually being done to find them.

The refactored code use LINQ to surface the intent of the code:
- Recursively find all the types declared under the HashFactory class
- Pick only types that are abstract sealed classes
- Instantiate instances of these algorithms
- Also add some Blake2 algorithms