Use Span<T> for CRC32 computations
The core of the CRC32 computations is marked as unsafe because it uses
pointers to try to efficiently process the blocks of memory. By
taking advantage of the newer Span<T> types, the need to use pointers
can be removed. Furthermore the code is made more legible with less
needs to take into account the current index and remaining lengths as
the data is scanned.

Also refactored the code to remove the two branches inside the tight
inner loop. The two branches checked the machine architecture to see
if the machine is little or big endian. Since this state will never
change at runtime (without a reboot), the check is done only once
outside the loop.