Further optimize use of Span<T>
Analyzed that the remained span could be easily computed based on the
number of blocks and there is no need for multiple allocations of
byte spans for it to settle into final position.

Further analyzed that computing spans of each 4 16-byte segment was
faster than trying to allocate spans of uint and walking that forward.

Using foreach() also seems to give the JIT a better chance of
optimizing the loop to process the large segments instead of using a
nested loop.