Improve Adler32 performance
The previous version would repeatedly compute an offset into a byte
array. Furthermore it would repeatedly try to determine whether it is
processing the maximum of 3800 bytes during the current iteration,
or if it finally at the last few bytes of data.

Changed the logic to handle all the big blocks first.

Modified the code to use a ReadOnlySpan<byte>. Although the
ReadOnlySpan<byte>.Enumerator also computes an offset into a byte array,
it looks to be performing better at about 1100 MB/s as compared to the
original code running at 685 MB/s.

For code running in 64-bit processes, data is read from memory 64-bits
at a time while working on the big blocks to get an additional 200 MB/s
more throughput bringing the speed up to 1300 MB/s.